style: move prompts to llm dir
